% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjure_commas.R
\name{conjure_commas}
\alias{conjure_commas}
\title{Conjure Commas}
\usage{
conjure_commas(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector of the same length as \code{x} with commas inserted as thousand separators.
}
\description{
Adds commas to large numbers for easier readability.
}
