% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inscribe.R
\name{inscribe}
\alias{inscribe}
\title{inscribe}
\usage{
inscribe(
  type = c("plot", "table"),
  github_username = "colebaril",
  bluesky_username = "@colebaril.ca"
)
}
\arguments{
\item{type}{Either "plot" or "table"}

\item{github_username}{Your GitHub username (default: "colebaril")}

\item{x}{A ggplot object or gt table}
}
\value{
A ggplot or gt object with a caption applied
}
\description{
Add a social media caption to a ggplot or gt table.
}
