% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spooky.R
\name{theme_spooky}
\alias{theme_spooky}
\title{theme_spooky()}
\usage{
theme_spooky()
}
\value{
A \code{ggplot2} theme object.
}
\description{
A \code{ggplot2} theme styled for Halloween with dark backgrounds,
glowing text, and spooky atmosphere.
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(color = "orange") +
  theme_spooky()

}
