% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_parchment.R
\name{theme_parchment}
\alias{theme_parchment}
\title{Theme Parchment}
\usage{
theme_parchment(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
A themed ggplot object
}
\description{
A ggplot2 theme styled like old parchment and ink.
}
